Modularizing rust stream
