+++ rust_stream/src/antd_client.rs
@@ -1,0 +1,62 @@
+use std::time::Duration;
+
+use base64::{engine::general_purpose::STANDARD as BASE64, Engine as _};
+use serde::Deserialize;
+
+#[derive(Clone)]
+pub(crate) struct AntdRestClient {
+    base_url: String,
+    client: reqwest::Client,
+}
+
+#[derive(Deserialize)]
+pub(crate) struct AntdHealthResponse {
+    pub(crate) status: String,
+    pub(crate) network: Option<String>,
+}
+
+#[derive(Deserialize)]
+struct AntdPublicDataResponse {
+    data: String,
+}
+
+impl AntdRestClient {
+    pub(crate) fn new(base_url: &str) -> anyhow::Result<Self> {
+        Ok(Self {
+            base_url: base_url.trim_end_matches('/').to_string(),
+            client: reqwest::Client::builder()
+                .connect_timeout(Duration::from_secs(5))
+                .timeout(Duration::from_secs(60))
+                .build()?,
+        })
+    }
+
+    pub(crate) async fn health(&self) -> anyhow::Result<AntdHealthResponse> {
+        self.get_json("/health").await
+    }
+
+    pub(crate) async fn data_get_public(&self, address: &str) -> anyhow::Result<Vec<u8>> {
+        let payload: AntdPublicDataResponse = self
+            .get_json(&format!("/v1/data/public/{}", address.trim()))
+            .await?;
+        BASE64
+            .decode(payload.data)
+            .map_err(|err| anyhow::anyhow!("antd returned invalid base64 public data: {err}"))
+    }
+
+    async fn get_json<T>(&self, path: &str) -> anyhow::Result<T>
+    where
+        T: for<'de> Deserialize<'de>,
+    {
+        let url = format!("{}{}", self.base_url, path);
+        let response = self.client.get(&url).send().await?;
+        let status = response.status();
+
+        if !status.is_success() {
+            let body = response.text().await.unwrap_or_else(|_| "".to_string());
+            anyhow::bail!("GET {path} failed: {status} {body}");
+        }
+
+        response.json::<T>().await.map_err(Into::into)
+    }
+}
